Rename ideal components
