+++ src/components/CMakeLists.txt
@@ -40,10 +40,10 @@ examples/example_serial_communication_for_hils.cpp
 examples/example_i2c_controller_for_hils.cpp
 examples/example_i2c_target_for_hils.cpp
 
-ideal_components/force_generator.cpp
-ideal_components/initialize_force_generator.cpp
-ideal_components/torque_generator.cpp
-ideal_components/initialize_torque_generator.cpp
+ideal/force_generator.cpp
+ideal/initialize_force_generator.cpp
+ideal/torque_generator.cpp
+ideal/initialize_torque_generator.cpp
 
 mission/telescope.cpp
 mission/initialize_telescope.cpp
