+++ src/components/ideal/initialize_force_generator.cpp
