+++ src/components/ideal/initialize_torque_generator.cpp
