+++ src/components/ideal/initialize_torque_generator.hpp
@@ -3,8 +3,8 @@
  * @brief Initialize function for TorqueGenerator
  */
 
-#ifndef S2E_COMPONENTS_IDEAL_COMPONENTS_INITIALIZE_TORQUE_GENERATOR_HPP_
-#define S2E_COMPONENTS_IDEAL_COMPONENTS_INITIALIZE_TORQUE_GENERATOR_HPP_
+#ifndef S2E_COMPONENTS_IDEAL_INITIALIZE_TORQUE_GENERATOR_HPP_
+#define S2E_COMPONENTS_IDEAL_INITIALIZE_TORQUE_GENERATOR_HPP_
 
 #include "torque_generator.hpp"
 
@@ -17,4 +17,4 @@
  */
 TorqueGenerator InitializeTorqueGenerator(ClockGenerator* clock_gen, const std::string file_name, const Dynamics* dynamics);
 
-#endif  // S2E_COMPONENTS_IDEAL_COMPONENTS_INITIALIZE_TORQUE_GENERATOR_HPP_
+#endif  // S2E_COMPONENTS_IDEAL_INITIALIZE_TORQUE_GENERATOR_HPP_
