+++ src/components/ideal/torque_generator.cpp
